change file names, add a tester script
also create initial mapping of accel + turn to motor outputs - will probably need redoing
