+++ main.py
@@ -1,8 +1,7 @@
-
-
+import gamepad_parser
 
 def main():
-
+    pass
 
 if __name__ == "__main__":
     main()
